Namespace refactor
