--- src/Animation/AnimationFrame.cs
@@ -1,10 +1,0 @@
-﻿using System; 
-
-using Microsoft.Xna.Framework;
-
-namespace Kazaam.Animate {
-  public class AnimationFrame {
-    public Rectangle SourceRectangle { get; set; }
-    public TimeSpan Duration { get; set; } 
-  }
-}
