+++ src/Audio/ISound.cs
