+++ src/Components/Collider.cs
@@ -1,5 +1,4 @@
-﻿using Kazaam.Physics;
-using Humper;
+﻿using Humper;
 using Microsoft.Xna.Framework;
 using System.Collections.Generic;
