+++ src/IPlayer.cs
@@ -4,7 +4,7 @@
 // Kazaam Simple Game Engine
 // Copyright (C) Connor Shlatz. All rights reserved.
 //-----------------------------------------------------------------------------
-using Kazaam.Input;
+using Kazaam;
 
 namespace Kazaam.Objects {
   public interface IPlayer {
