+++ src/Loaders/IContentLoader.cs
@@ -1,6 +1,6 @@
 using Microsoft.Xna.Framework.Content;
 
-namespace Kazaam.Assets {
+namespace Kazaam.Loaders {
     public interface IContentLoader {
         ContentManager Content { get; set; }
         object Load(string contentPath);
