+++ src/Loaders/SongLoader.cs
@@ -1,7 +1,7 @@
 using Microsoft.Xna.Framework.Content;
 using Microsoft.Xna.Framework.Media;
 
-namespace Kazaam.Assets {
+namespace Kazaam.Loaders {
     public class SongLoader : IContentLoader {
         public ContentManager Content { get; set; }
         public object Load(string contentPath) {
